feat: ensure cart product quantity can be decreased

+++ src/actions/decrease-cart-product-quantity/index.ts
@@ -8,12 +8,12 @@ import { db } from "@/db";
 import { cartItemTable } from "@/db/schema";
 import { auth } from "@/lib/auth";
 
-import { removeProductFromCartSchema } from "./schema";
+import { decreaseCartProductQuantitySchema } from "./schema";
 
-export async function removeProductFromCart(
-  data: z.infer<typeof removeProductFromCartSchema>,
-) {
-  removeProductFromCartSchema.parse(data);
+export const decreaseCartProductQuantity = async (
+  data: z.infer<typeof decreaseCartProductQuantitySchema>,
+) => {
+  decreaseCartProductQuantitySchema.parse(data);
 
   const session = await auth.api.getSession({
     headers: await headers(),
@@ -33,12 +33,19 @@ export async function removeProductFromCart(
   if (!cartItem) {
     throw new Error("Cart item not found");
   }
-
-  const cartDoesNotBelongToUser = cartItem?.cart.userId !== session.user.id;
+  const cartDoesNotBelongToUser = cartItem.cart.userId !== session.user.id;
 
   if (cartDoesNotBelongToUser) {
     throw new Error("Unauthorized");
   }
 
-  await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+  if (cartItem.quantity === 1) {
+    await db.delete(cartItemTable).where(eq(cartItemTable.id, cartItem.id));
+    return;
+  }
+
+  await db
+    .update(cartItemTable)
+    .set({ quantity: cartItem.quantity - 1 })
+    .where(eq(cartItemTable.id, cartItem.id));
 };
